pyproject.toml
[tool.poetry]
name = "tach-challenge"
version = "0.1.0"
description = ""
authors = ["Lucas Lucyk <llucyk@grupospec.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "petisco_sanic"}, {include = "accounts"}, {include = "transactions"}]

[tool.poetry.dependencies]
python = "^3.10"
sanic = {extras = ["ext"], version = "^23.12.1"}
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
aiormq = "^6.8.0"
beanie = "^1.25.0"
python-decouple = "^3.8"
meiga = "^1.9.7"
petisco = {extras = ["rabbitmq"], version = "^2.1.3"}
anyio = "^4.3.0"


[tool.poetry.group.dev.dependencies]
diagrams = "^0.23.4"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
